General code re-organization
